Refactored common logic between the two webhook service classes into BaseWebhook.  The connection setup and work url generation were exactly the same.
