+++ app/services/webhooks/work_published_webhook.rb
@@ -1,25 +1,15 @@
 # frozen_string_literal: true
 
 module Webhooks
-  class WorkPublishedWebhook
-    def initialize(work_uuid)
-      @work_uuid = work_uuid
-    end
-
-    def notify
-      return if ENV['RMD_HOST'].blank?
-
-      conn = Faraday.new(
-        url: ENV['RMD_HOST'],
-        headers: { 'X-API-KEY' => ENV['RMD_WEBHOOK_SECRET'] }
-      )
-
-      conn.post('/webhooks/scholarsphere/open_access_work_published',
-                scholarsphere_work_url: "https://#{Rails.application.routes.default_url_options[:host]}/resources/#{work_uuid}")
-    end
-
+  class WorkPublishedWebhook < BaseWebhook
     private
 
-      attr_reader :work_uuid
+      def webhook_path
+        '/webhooks/scholarsphere/open_access_work_published'
+      end
+
+      def payload_key
+        :scholarsphere_work_url
+      end
   end
 end
