+++ spec/services/webhooks/work_removed_webhook_spec.rb
@@ -9,11 +9,14 @@ RSpec.describe Webhooks::WorkRemovedWebhook do
     let(:faraday_connection) { instance_spy Faraday::Connection }
 
     before do
+      ENV['RMD_HOST'] = 'https://rmd.example.com'
       allow(Faraday).to receive(:new).with(
         url: ENV['RMD_HOST'],
         headers: { 'X-API-KEY' => ENV['RMD_WEBHOOK_SECRET'] }
       ).and_return faraday_connection
     end
+
+    after { ENV.delete('RMD_HOST') }
 
     it 'posts the URL for the work to the RMD webhook endpoint' do
       webhook.notify
